Frozen Cortex baseline GL code
Squashed commit of the following:

commit e2ce6acc7e4a8b61a31b96335942725123fbbf40
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 20:53:00 2014 +0100

    More stuff to full

commit 8183d315dd08b5bd61a85a640498cede2246f407
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 20:50:15 2014 +0100

    Fix copyright merge fail

commit af2f118ab89c68822d2347b1e393310c1675fb69
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 18:34:34 2014 +0100

    Add new shaders for empty

commit dff98653a8d3961d235c52bba91bc17c5aaa5ec4
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 18:33:24 2014 +0100

    Re-fix engine

commit e2749d18afbdc28ea46948929cc2bcd29a01de91
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 18:33:12 2014 +0100

    Re-fix game files

commit 26adbcc359e8f94af58569ae72a2e6c6aec5d546
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 17:47:29 2014 +0100

    Add glew to linux project

commit 2374309c27f0918017ce1b1644f5a658e602b00a
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 17:27:01 2014 +0100

    Add glew

commit 8412931c2fab66412ae77deb603388a8894d16b8
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 17:18:55 2014 +0100

    Update copyright notices in files

commit 6b28fa8ea3a149530b2fd6e377dda8d2e91207c6
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 17:18:14 2014 +0100

    Don't need this

commit cdec717b53551e0ea2d5feab63e12d0b1fd9cf60
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 14:51:03 2014 +0100

    Add back in mesh names for debugging render states

commit e3797fa21273f9e10c22772500c356958db79c14
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 14:50:24 2014 +0100

    Fix mac platform

commit ef6e7a18a64b1dd46d564deecfd32461d6eca865
Author: James Urquhart <jamesu@gmail.com>
Date:   Fri Sep 12 14:43:54 2014 +0100

    Fix Terrain and Advanced Lighting Effects

commit 077db6f76595490ebea9769aa6252b2db916ed21
Author: James Urquhart <jamesu@gmail.com>
Date:   Thu Sep 11 20:57:12 2014 +0100

    Basic GL 2.x branch code

+++ Engine/source/windowManager/mac/macCursorController.mm
@@ -31,7 +31,7 @@ void MacCursorController::setCursorPosition(S32 x, S32 y)
    if(!macWindow || !macWindow->isVisible())
       return;
       
-   CGPoint pt = { x, y };
+   CGPoint pt = { (CGFloat)x, (CGFloat)y };
    CGWarpMouseCursorPosition(pt);
    
    macWindow->_skipAnotherMouseEvent();
